Separated `gen/register.go` into multiple files
