+++ gen/register_generator.go
@@ -6,48 +6,6 @@ import (
 	"alon.kr/x/usm/parse"
 )
 
-// MARK: Info
-type RegisterInfo struct {
-	// The name of the register, as it appears in the source code.
-	Name string
-
-	// The type of the register.
-	Type ReferencedTypeInfo
-
-	// The first location in the source code in which the register is declared
-	// or assigned a value.
-	Declaration core.UnmanagedSourceView
-}
-
-func (i RegisterInfo) toPartialRegisterInfo() partialRegisterInfo {
-	return partialRegisterInfo{
-		Name:        i.Name,
-		Type:        &i.Type,
-		Declaration: i.Declaration,
-	}
-}
-
-// This represents partial register information, possibly without an associated
-// type (yet). This is used internally before the compiler has finally determined
-// the type of the register, if the type is implicit.
-type partialRegisterInfo struct {
-	Name string
-
-	// Possibly nil, if type is implicitly defined.
-	Type *ReferencedTypeInfo
-
-	Declaration core.UnmanagedSourceView
-}
-
-// MARK: Manager
-
-type RegisterManager interface {
-	GetRegister(name string) *RegisterInfo
-	NewRegister(reg *RegisterInfo) core.Result
-}
-
-// MARK: Generator
-
 // Used to convert parse.RegisterNode instances to *existing* register instances.
 // Returns an error on generation if the provided register node references a
 // register that does not exist.
