+++ gen/target.go
@@ -12,8 +12,8 @@ type TargetGenerator[InstT BaseInstruction] struct {
 	ReferencedTypeGenerator FileContextGenerator[InstT, parse.TypeNode, ReferencedTypeInfo]
 }
 
-func NewTargetGenerator[InstT BaseInstruction]() FunctionContextGenerator[InstT, parse.TargetNode, partialRegisterInfo] {
-	return FunctionContextGenerator[InstT, parse.TargetNode, partialRegisterInfo](
+func NewTargetGenerator[InstT BaseInstruction]() FunctionContextGenerator[InstT, parse.TargetNode, registerPartialInfo] {
+	return FunctionContextGenerator[InstT, parse.TargetNode, registerPartialInfo](
 		&TargetGenerator[InstT]{
 			ReferencedTypeGenerator: NewReferencedTypeGenerator[InstT](),
 		},
@@ -41,7 +41,7 @@ func NewRegisterTypeMismatchResult(
 func (g *TargetGenerator[InstT]) Generate(
 	ctx *FunctionGenerationContext[InstT],
 	node parse.TargetNode,
-) (partialRegisterInfo, core.ResultList) {
+) (registerPartialInfo, core.ResultList) {
 	var explicitType *ReferencedTypeInfo
 
 	// if an explicit type is provided to the target, get the type info.
@@ -52,7 +52,7 @@ func (g *TargetGenerator[InstT]) Generate(
 			*node.Type,
 		)
 		if !results.IsEmpty() {
-			return partialRegisterInfo{}, results
+			return registerPartialInfo{}, results
 		}
 
 		explicitType = &explicitTypeValue
@@ -66,7 +66,7 @@ func (g *TargetGenerator[InstT]) Generate(
 		if explicitType != nil {
 			// ensure explicit type matches the previously declared one.
 			if !explicitType.Equal(registerInfo.Type) {
-				return partialRegisterInfo{}, NewRegisterTypeMismatchResult(
+				return registerPartialInfo{}, NewRegisterTypeMismatchResult(
 					node.View(),
 					registerInfo.Declaration,
 				)
@@ -82,7 +82,7 @@ func (g *TargetGenerator[InstT]) Generate(
 		// the target register at this.
 		// the type and register will be finalized when the instruction is built,
 		// and only then it is added to the register manager.
-		return partialRegisterInfo{
+		return registerPartialInfo{
 			Name:        registerName,
 			Type:        explicitType,
 			Declaration: node.View(),
